He arreglado que antes con el metodo DELETE no borraba el contacto

+++ src/main/java/com/Benjamin/repositorios/ContactoRepository.java
@@ -34,7 +34,7 @@ public class ContactoRepository {
         return contacto;
     }
 
-    public boolean eliminar(Long id) {
+    public boolean eliminar(int id) {
         for (int i = 0; i < contactos.size(); i++) {
             if (contactos.get(i).getId() == id) {
                 contactos.remove(i);
